move getSecurityCodePrefix to contractkit
