+++ packages/mobile/src/identity/securityCode.test.ts
@@ -1,12 +1,4 @@
-import { extractSecurityCodeWithPrefix, getSecurityCodePrefix } from 'src/identity/securityCode'
-
-describe(getSecurityCodePrefix, () => {
-  it('should compute correct hash', () => {
-    expect(getSecurityCodePrefix('0x000000000000000000000008')).toEqual('8')
-    // 0xf7f551752A78Ce650385B58364225e5ec18D96cB -> 1415591498931780605110544902041322891412830525131
-    expect(getSecurityCodePrefix('0xf7f551752A78Ce650385B58364225e5ec18D96cB')).toEqual('1')
-  })
-})
+import { extractSecurityCodeWithPrefix } from 'src/identity/securityCode'
 
 describe(extractSecurityCodeWithPrefix, () => {
   it('should extract 8 digit code', () => {
